requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
astunparse==1.6.3
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cmake==4.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.2.4
deepface==0.0.97
easydict==1.13
filelock==3.20.3
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.2
Flask-Mail==0.10.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
gast==0.4.0
gdown==5.2.1
google-auth==2.45.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.76.0
gunicorn==23.0.0
h5py==3.15.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
insightface==0.7.3
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
joblib==1.5.3
keras==2.12.0
keras-facenet==0.3.2
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
lightecc==0.0.4
lightphe==0.0.20
lz4==4.4.5
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
mysql-connector-python==9.5.0
networkx==3.4.2
numpy==1.23.5
oauthlib==3.3.1
onnx==1.20.1
onnxruntime==1.23.2
opencv-python==4.8.1.78
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.3
pillow==10.2.0
prettytable==3.17.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
pyreadline3==3.5.4
PySimpleGUI==5.0.8.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
retina-face==0.0.17
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
simsimd==6.5.12
six==1.17.0
soupsieve==2.8.1
stringzilla==4.6.0
sympy==1.14.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tk==0.1.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wcwidth==0.2.14
Werkzeug==3.1.4
wrapt==1.14.2
